+++ CHANGELOG.rst
@@ -60,6 +60,22 @@ ProPlot v0.3.0 (2020-01-##)
   with a default ``.proplotrc`` file, change the auto-generated user ``.proplotrc``
   (:pr:`50`).
 
+ProPlot v0.2.8 (2019-12-09)
+===========================
+.. rubric:: Deprecated
+
+- Remove ``'Moisture'`` colormap (:commit:`cf8952b1`).
+
+.. rubric:: Features
+
+- New ``'DryWet'`` colormap is colorblind friendly (:commit:`0280e266`).
+
+.. rubric:: Bug fixes
+
+- Add brute force workaround for saving colormaps with
+  *callable* segmentdata (:commit:`8201a806`).
+- Save the ``cyclic`` and ``gamma`` attributes in JSON files too (:commit:`8201a806`).
+
 ProPlot v0.2.7 (2019-12-09)
 ===========================